chore: 게임팀플레이어 로직들 분리 #75
