+++ game/application/__init__.py
@@ -1,3 +1,4 @@
 from .game_serivice import GameService
 from .game_team_serivce import GameTeamService
 from .game_team_get_service import GameTeamGetService
+from .game_team_player_service import GameTeamPlayerService
